% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSWAT2009.R
\name{runSWAT2009}
\alias{runSWAT2009}
\title{This runs the SWAT2009 executable in the current directory.}
\usage{
runSWAT2009(hist_wx = NULL, elev = 100, rch = 3)
}
\arguments{
\item{hist_wx}{Describe \code{hist_wx}}

\item{elev}{Describe \code{elev}}

\item{rch}{Describe \code{rch}}
}
\description{
This function runs the SWAT2009 executable in the current directory.
}
\author{
Daniel R. Fuka
}
