% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatr-package.R
\docType{data}
\name{swat_general}
\alias{swat_general}
\title{All files required for a general SWAT run.}
\description{
All files required for a general SWAT run. This initialization has 3
subbasin init, with 3 HRUs per subbasin.
}
\examples{

data(swat_general)

}
